Backend/requirements.txt
Django==5.2.8
djangorestframework==3.15.2
django-cors-headers==4.6.0
djangorestframework-simplejwt==5.4.0
Pillow==11.0.0
python-decouple==3.8
